# sdks/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: aju API
  version: 0.1.0
  summary: HTTP API for the aju memory platform.
  description: |
    The aju HTTP API. This spec is the single source of truth for every
    client SDK (TypeScript, Python, Go). Changes here flow into all SDKs
    via `sdks/scripts/generate.sh`.

    Scope (v0.1): the vault content API and brain listing — the surface
    that third-party integrations and agents need. Auth, signup, OAuth,
    org administration, and session-only routes are intentionally
    excluded from the SDK surface.
  license:
    name: See repository LICENSE
servers:
  - url: https://aju.sh
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/vault/deep-search:
    get:
      tags: [vault]
      operationId: deepSearchVault
      summary: Graph-RAG search — follows wikilinks from seed documents.
      parameters:
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/BrainsQuery"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/vault/create:
    post:
      tags: [vault]
      operationId: createDocument
      summary: Create a document.
      parameters:
        - $ref: "#/components/parameters/BrainQuery"
        - name: defer_index
          in: query
          description: Skip per-create link rebuild + embedding. Caller should invoke `reindex` after bulk import.
          schema: { type: string, enum: ["1"] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentWrite" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/vault/update:
    post:
      tags: [vault]
      operationId: updateDocument
      summary: Update a document in place (by path).
      parameters:
        - $ref: "#/components/parameters/BrainQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentWrite" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/vault/rebuild-links:
    post:
      tags: [vault]
      operationId: rebuildLinks
      summary: Rebuild only the wikilink graph for a brain.
      parameters:
        - $ref: "#/components/parameters/BrainQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  linksBuilt: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key (aju_live_… or aju_test_…)
  parameters:
...
    BrainQuery:
      name: brain
      in: query
      description: Brain name. Omit to target the caller's default brain.
      schema: { type: string }
    BrainsQuery:
      name: brain
      in: query
      description: |
        Target brain(s). Accepts a single name, comma-separated list,
        repeated query param, or the literal "all". Omit for the default.
      schema: { type: string }
    Query:
      name: q
      in: query
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not authorized for this brain or action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: { type: string }
...
    DocumentWrite:
      type: object
      required: [path, content, source]
      properties:
        path:
          type: string
          description: Vault path, e.g. "journal/2026-04-24.md".
        content:
          type: string
          description: Markdown body including optional YAML frontmatter.
        source:
          type: string
          description: Free-form label for the origin of this write (e.g. "cli", "sdk-ts").
    Document:
      type: object
      required: [id, path, title, brainId]
      properties:
        id: { type: string }
        brainId: { type: string }
        path: { type: string }
        title: { type: string }
        content: { type: string }
        contentHash: { type: string }
        docType: { type: string, nullable: true }
        docStatus: { type: string, nullable: true }
        section: { type: string, nullable: true }
        directory: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        wikilinks:
          type: array
          items: { type: string }
        frontmatter:
          type: object
          additionalProperties: true
          nullable: true
        wordCount: { type: integer }
        fileModified: { type: string, format: date-time }
        syncedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    SearchResult:
      type: object
      required: [id, path, title]
      properties:
        id: { type: string }
        path: { type: string }
        title: { type: string }
        section: { type: string, nullable: true }
        docType: { type: string, nullable: true }
        docStatus: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
        wordCount: { type: integer }
        sourceType:
          type: string
          enum: [document, file]
        mimeType: { type: string, nullable: true }
        brain: { type: string, nullable: true }
        rank: { type: number, format: float }
        snippet: { type: string }
    SearchResponse:
      type: object
      required: [query, count, results]
      properties:
        query: { type: string }
        brains:
          type: array
          items: { type: string }
        count: { type: integer }
        results:
          type: array
          items: { $ref: "#/components/schemas/SearchResult" }
...